Changed pointerId from `Int` to `Double` > https://www.w3.org/TR/pointerevents/#pointerevent-interface says that pointerId is a long in the IDL, so it should be a Double in Scala.js, not an Int (and not a Long either).

+++ src/main/scala/org/scalajs/dom/raw/lib.scala
@@ -1478,7 +1478,7 @@ class PointerEvent extends MouseEvent {
    *
    * MDN
    */
-  def pointerId: Int = js.native
+  def pointerId: Double = js.native
 
   /**
    * The width (magnitude on the X axis), in CSS pixels, of the contact geometry of the pointer.
